# backend/pkg/swagger/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a mall-tenants API",
        "title": "mall-tenants",
        "contact": {
            "name": "Dmitry Boyarkin",
            "email": "boyarkin_dima2@mail.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/payment/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новый платеж для аренды",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Создать платеж",
                "parameters": [
                    {
                        "description": "Данные платежа",
                        "name": "payment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Payment created successfully"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/rental/{rentalID}/last": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает последний платеж по rentalID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Получить последний платеж",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID аренды",
                        "name": "rentalID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/rental/{rentalID}/overdue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список просроченных платежей по rentalID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Получить просроченные платежи",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID аренды",
                        "name": "rentalID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Payment"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment/{paymentID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает информацию о платеже по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payment"
                ],
                "summary": "Получить платеж",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID платежа",
                        "name": "paymentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Payment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/premise": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список помещений",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Получить список помещений",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Premises"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Помещений не найдено",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/premise/": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создаёт новое помещение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Создать помещение",
                "parameters": [
                    {
                        "description": "Информация о помещении",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Premises"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет данные о помещении",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Обновить помещение",
                "parameters": [
                    {
                        "description": "Информация о помещении",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Premises"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Помещение не найдено",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/premise/{code}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает данные о помещении по коду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "premises"
                ],
                "summary": "Получить данные о помещении",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Код помещения",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Premises"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Помещение не найдено",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/sys.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "payment_date": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "integer"
                },
                "rent_id": {
                    "type": "integer"
                }
            }
        },
...
        "model.Premises": {
            "type": "object",
            "properties": {
                "air_conditioning": {
                    "type": "string"
                },
                "area": {
                    "type": "integer"
                },
                "code": {
                    "type": "integer"
                },
                "floor": {
                    "type": "integer"
                },
                "rent_per_month": {
                    "type": "integer"
                },
                "security_system": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/model.PremisesType"
                }
            }
        },
        "model.PremisesType": {
            "type": "string",
            "enum": [
                "retail",
                "service"
            ],
            "x-enum-varnames": [
                "Retail",
                "Service"
            ]
        },
...
        "sys.ErrorResponse": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}